refactor(ui): App.vue をコンポーネントに分割し @ エイリアスを導入する
- TabGroup / DayNav / ThemeToggle / MeterRing / DayStamps を src/components/ に切り出す
- vite.config.ts と tsconfig.json に @ → src/ エイリアスを設定
- App.vue はデータ取得・日付計算・レイアウト組み立てのみに整理
