+++ src/components/ThemeToggle.vue
@@ -1,0 +1,48 @@
+<script setup lang="ts">
+import { Sun } from 'lucide-vue-next';
+</script>
+
+<template>
+  <button class="theme-toggle" aria-label="テーマ切り替え">
+    <span class="toggle-track">
+      <span class="toggle-thumb">
+        <Sun :size="12" />
+      </span>
+    </span>
+  </button>
+</template>
+
+<style scoped>
+.theme-toggle {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 54px;
+  height: 44px;
+  border: none;
+  background: transparent;
+  cursor: pointer;
+  padding: 0;
+}
+
+.toggle-track {
+  display: flex;
+  align-items: center;
+  width: 54px;
+  height: 28px;
+  background-color: var(--sub-color);
+  border-radius: 14px;
+  padding: 2px;
+}
+
+.toggle-thumb {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 24px;
+  height: 24px;
+  background-color: var(--pond-color);
+  border-radius: 50%;
+  color: var(--base-color);
+}
+</style>
